refactor(provider-registry): fold endpoint dispatch into registry-utils
`utils/modelRouting.ts` held two functions that read a provider's endpoint
configs and answer which endpoint a model uses — the same job as the
`registry-utils.ts` neighbours that interpret registry data
(`buildPersistedEndpointConfigs`, `inferAdapterFamily`,
`endpointImpliedCapability`). It is also `buildPersistedEndpointConfigs` that
carries `serves` out to the runtime, so the pair now sits in one file.

`resolveModelEndpoint` stays a function rather than being inlined: it encodes
the declared → claim → default priority for two callers — the request path and
the catalog's reasoning projection — and those two disagreeing is what #17900
was. It now also returns the resolved namespace instead of the raw route, so the
"only use the claim's namespace if the claim actually won" check lives with the
priority chain it depends on rather than being restated by each caller.
