+++ src/main/ai/provider/endpoint.ts
@@ -44,15 +44,14 @@ export function resolveWireModelId(model: Model, endpointType: EndpointType | un
  * fallback among `endpointConfigs`.
  */
 export function resolveEffectiveEndpoint(provider: Provider, model: Model): ResolvedEndpoint {
-  // A route can only come from an endpoint the provider declares, so an incomplete row cannot
-  // route to a backend that isn't there — the claim and the connection config are the same object.
-  const { endpointType, route } = resolveModelEndpoint({
+  // A claim can only come from an endpoint the provider declares, so an incomplete row cannot route
+  // to a backend that isn't there — the claim and the connection config are the same object.
+  const { endpointType, providerOptionsKey } = resolveModelEndpoint({
     endpointTypes: model.endpointTypes,
     endpointConfigs: provider.endpointConfigs,
     modelId: model.apiModelId ?? model.id,
     defaultChatEndpoint: provider.defaultChatEndpoint
   })
-  const providerOptionsKey = route && endpointType === route.endpointType ? route.providerOptionsKey : undefined
   return { endpointType, baseUrl: getBaseUrl(provider, endpointType), providerOptionsKey }
 }
